+++ packages/core/src/models/actor.ts
