+++ packages/core/src/models/index.ts
@@ -1,1 +1,5 @@
+export * as Actor from './actor.js'
+export * as Curriculum from './curriculum/index.js'
+export * as Location from './location/index.js'
+export * as Program from './program/index.js'
 export * from './programs.js'
