+++ packages/core/src/models/location/index.ts
@@ -2,14 +2,14 @@ import { schema } from '@nawadi/db'
 import { eq } from 'drizzle-orm'
 import { createSelectSchema } from 'drizzle-zod'
 import { z } from 'zod'
-import { useTransaction } from '../util/transaction.js'
-import { zod } from '../util/zod.js'
+import { useTransaction } from '../../util/transaction.js'
+import { zod } from '../../util/zod.js'
 
-export * as Module from './module.js'
+export * as Location from './index.js'
 
-const module = schema.module
+const location = schema.location
 
-export const Info = createSelectSchema(module, {
+export const Info = createSelectSchema(location, {
   handle(schema) {
     return schema.handle
       .trim()
@@ -17,25 +17,33 @@ export const Info = createSelectSchema(module, {
       .min(3)
       .regex(/^[a-z0-9\-]+$/)
   },
+  logoMediaId: (schema) => schema.logoMediaId.uuid(),
+  squareLogoMediaId: (schema) => schema.squareLogoMediaId.uuid(),
+  websiteUrl: (schema) => schema.websiteUrl.url(),
 })
-export type Info = typeof module.$inferSelect
+export type Info = typeof location.$inferSelect
 
 export const create = zod(
-  Info.pick({ title: true, handle: true }).partial({
+  Info.pick({
     title: true,
+    handle: true,
+    name: true,
+    websiteUrl: true,
+  }).partial({
+    name: true,
+    websiteUrl: true,
   }),
   (input) =>
     useTransaction(async (tx) => {
       const [insert] = await tx
-        .insert(module)
+        .insert(location)
         .values({
-          handle: input.handle,
-          title: input.title,
+          ...input,
         })
-        .returning({ id: module.id })
+        .returning({ id: location.id })
 
       if (!insert) {
-        throw new Error('Failed to insert module')
+        throw new Error('Failed to insert location')
       }
 
       return insert.id
@@ -44,7 +52,7 @@ export const create = zod(
 
 export const list = zod(z.void(), async () =>
   useTransaction(async (tx) => {
-    return tx.select().from(module)
+    return tx.select().from(location)
   }),
 )
 
@@ -52,8 +60,8 @@ export const fromId = zod(Info.shape.id, async (id) =>
   useTransaction(async (tx) => {
     return tx
       .select()
-      .from(module)
-      .where(eq(module.id, id))
+      .from(location)
+      .where(eq(location.id, id))
       .then((rows) => rows[0])
   }),
 )
@@ -62,8 +70,8 @@ export const fromHandle = zod(Info.shape.handle, async (handle) =>
   useTransaction(async (tx) => {
     return tx
       .select()
-      .from(module)
-      .where(eq(module.handle, handle))
+      .from(location)
+      .where(eq(location.handle, handle))
       .then((rows) => rows[0])
   }),
 )
