+++ packages/core/src/models/program/category.ts
@@ -2,12 +2,16 @@ import { schema } from '@nawadi/db'
 import { eq } from 'drizzle-orm'
 import { createSelectSchema } from 'drizzle-zod'
 import { z } from 'zod'
-import { useTransaction } from '../util/transaction.js'
-import { zod } from '../util/zod.js'
+import { useTransaction } from '../../util/transaction.js'
+import { zod } from '../../util/zod.js'
 
-export * as Competency from './competency.js'
+export * as Category from './category.js'
 
-export const Info = createSelectSchema(schema.competency, {
+const category = schema.category
+
+export const test = 1
+
+export const Info = createSelectSchema(category, {
   handle(schema) {
     return schema.handle
       .trim()
@@ -15,26 +19,35 @@ export const Info = createSelectSchema(schema.competency, {
       .min(3)
       .regex(/^[a-z0-9\-]+$/)
   },
+  parentCategoryId: (schema) => schema.parentCategoryId.uuid(),
 })
-export type Info = typeof schema.competency.$inferSelect
+export type Info = typeof category.$inferSelect
 
 export const create = zod(
-  Info.pick({ title: true, type: true, handle: true }).partial({
+  Info.pick({
     title: true,
+    handle: true,
+    parentCategoryId: true,
+    description: true,
+  }).partial({
+    title: true,
+    parentCategoryId: true,
+    description: true,
   }),
   (input) =>
     useTransaction(async (tx) => {
       const [insert] = await tx
-        .insert(schema.competency)
+        .insert(category)
         .values({
           handle: input.handle,
-          type: input.type,
           title: input.title,
+          parentCategoryId: input.parentCategoryId,
+          description: input.description,
         })
-        .returning({ id: schema.competency.id })
+        .returning({ id: category.id })
 
       if (!insert) {
-        throw new Error('Failed to insert competency')
+        throw new Error('Failed to insert category')
       }
 
       return insert.id
@@ -43,7 +56,7 @@ export const create = zod(
 
 export const list = zod(z.void(), async () =>
   useTransaction(async (tx) => {
-    return tx.select().from(schema.competency)
+    return tx.select().from(category)
   }),
 )
 
@@ -51,8 +64,8 @@ export const fromId = zod(Info.shape.id, async (id) =>
   useTransaction(async (tx) => {
     return tx
       .select()
-      .from(schema.competency)
-      .where(eq(schema.competency.id, id))
+      .from(category)
+      .where(eq(category.id, id))
       .then((rows) => rows[0])
   }),
 )
@@ -61,8 +74,8 @@ export const fromHandle = zod(Info.shape.handle, async (handle) =>
   useTransaction(async (tx) => {
     return tx
       .select()
-      .from(schema.competency)
-      .where(eq(schema.competency.handle, handle))
+      .from(category)
+      .where(eq(category.handle, handle))
       .then((rows) => rows[0])
   }),
 )
