+++ packages/core/src/models/program/competency.ts
@@ -2,14 +2,12 @@ import { schema } from '@nawadi/db'
 import { eq } from 'drizzle-orm'
 import { createSelectSchema } from 'drizzle-zod'
 import { z } from 'zod'
-import { useTransaction } from '../util/transaction.js'
-import { zod } from '../util/zod.js'
+import { useTransaction } from '../../util/transaction.js'
+import { zod } from '../../util/zod.js'
 
-export * as Degree from './degree.js'
+export * as Competency from './competency.js'
 
-const degree = schema.degree
-
-export const Info = createSelectSchema(degree, {
+export const Info = createSelectSchema(schema.competency, {
   handle(schema) {
     return schema.handle
       .trim()
@@ -17,29 +15,26 @@ export const Info = createSelectSchema(degree, {
       .min(3)
       .regex(/^[a-z0-9\-]+$/)
   },
-  rang(schema) {
-    return schema.rang.int()
-  },
 })
-export type Info = typeof degree.$inferSelect
+export type Info = typeof schema.competency.$inferSelect
 
 export const create = zod(
-  Info.pick({ title: true, handle: true, rang: true }).partial({
+  Info.pick({ title: true, type: true, handle: true }).partial({
     title: true,
   }),
   (input) =>
     useTransaction(async (tx) => {
       const [insert] = await tx
-        .insert(degree)
+        .insert(schema.competency)
         .values({
           handle: input.handle,
+          type: input.type,
           title: input.title,
-          rang: input.rang,
         })
-        .returning({ id: degree.id })
+        .returning({ id: schema.competency.id })
 
       if (!insert) {
-        throw new Error('Failed to insert degree')
+        throw new Error('Failed to insert competency')
       }
 
       return insert.id
@@ -48,7 +43,7 @@ export const create = zod(
 
 export const list = zod(z.void(), async () =>
   useTransaction(async (tx) => {
-    return tx.select().from(degree)
+    return tx.select().from(schema.competency)
   }),
 )
 
@@ -56,8 +51,8 @@ export const fromId = zod(Info.shape.id, async (id) =>
   useTransaction(async (tx) => {
     return tx
       .select()
-      .from(degree)
-      .where(eq(degree.id, id))
+      .from(schema.competency)
+      .where(eq(schema.competency.id, id))
       .then((rows) => rows[0])
   }),
 )
@@ -66,8 +61,8 @@ export const fromHandle = zod(Info.shape.handle, async (handle) =>
   useTransaction(async (tx) => {
     return tx
       .select()
-      .from(degree)
-      .where(eq(degree.handle, handle))
+      .from(schema.competency)
+      .where(eq(schema.competency.handle, handle))
       .then((rows) => rows[0])
   }),
 )
