+++ packages/core/src/models/program/module.ts
@@ -2,14 +2,14 @@ import { schema } from '@nawadi/db'
 import { eq } from 'drizzle-orm'
 import { createSelectSchema } from 'drizzle-zod'
 import { z } from 'zod'
-import { useTransaction } from '../util/transaction.js'
-import { zod } from '../util/zod.js'
+import { useTransaction } from '../../util/transaction.js'
+import { zod } from '../../util/zod.js'
 
-export * as Location from './index.js'
+export * as Module from './module.js'
 
-const location = schema.location
+const module = schema.module
 
-export const Info = createSelectSchema(location, {
+export const Info = createSelectSchema(module, {
   handle(schema) {
     return schema.handle
       .trim()
@@ -17,33 +17,25 @@ export const Info = createSelectSchema(location, {
       .min(3)
       .regex(/^[a-z0-9\-]+$/)
   },
-  logoMediaId: (schema) => schema.logoMediaId.uuid(),
-  squareLogoMediaId: (schema) => schema.squareLogoMediaId.uuid(),
-  websiteUrl: (schema) => schema.websiteUrl.url(),
 })
-export type Info = typeof location.$inferSelect
+export type Info = typeof module.$inferSelect
 
 export const create = zod(
-  Info.pick({
+  Info.pick({ title: true, handle: true }).partial({
     title: true,
-    handle: true,
-    name: true,
-    websiteUrl: true,
-  }).partial({
-    name: true,
-    websiteUrl: true,
   }),
   (input) =>
     useTransaction(async (tx) => {
       const [insert] = await tx
-        .insert(location)
+        .insert(module)
         .values({
-          ...input,
+          handle: input.handle,
+          title: input.title,
         })
-        .returning({ id: location.id })
+        .returning({ id: module.id })
 
       if (!insert) {
-        throw new Error('Failed to insert location')
+        throw new Error('Failed to insert module')
       }
 
       return insert.id
@@ -52,7 +44,7 @@ export const create = zod(
 
 export const list = zod(z.void(), async () =>
   useTransaction(async (tx) => {
-    return tx.select().from(location)
+    return tx.select().from(module)
   }),
 )
 
@@ -60,8 +52,8 @@ export const fromId = zod(Info.shape.id, async (id) =>
   useTransaction(async (tx) => {
     return tx
       .select()
-      .from(location)
-      .where(eq(location.id, id))
+      .from(module)
+      .where(eq(module.id, id))
       .then((rows) => rows[0])
   }),
 )
@@ -70,8 +62,8 @@ export const fromHandle = zod(Info.shape.handle, async (handle) =>
   useTransaction(async (tx) => {
     return tx
       .select()
-      .from(location)
-      .where(eq(location.handle, handle))
+      .from(module)
+      .where(eq(module.handle, handle))
       .then((rows) => rows[0])
   }),
 )
